decopule ContainerFactory
